// front/src/app/utils/enums.ts
export enum ThemeColor {
  LIGHT = 'light',
  DARK = 'dark',
}

export enum StatusColor {
  Success = 'success',
  Info = 'info',
  Warning = 'warning',
  Danger = 'danger',
}

export enum FunctionName {
  AddPauser = 'AddPauser',
  AddVerified = 'AddVerified',
  Allowance = 'Allowance',
  Approve = 'Approve',
  ApproveAndCall = 'ApproveAndCall',
  AuthorizeOperator = 'AuthorizeOperator',
  BalanceOf = 'BalanceOf',
  BalanceOfID = 'BalanceOfID',
  BalanceOfBatch = 'BalanceOfBatch',
  Burn = 'Burn',
  BurnData = 'BurnData',
  BurnFrom = 'BurnFrom',
  CancelAndReissue = 'CancelAndReissue',
  ImplementInterfaceForAddress = 'ImplementInterfaceForAddress',
  Cap = 'Cap',
  ChangeOwner = 'ChangeOwner',
  CIDByHash = 'CIDByHash',
  Decimals = 'Decimals',
  DecreaseAllowance = 'DecreaseAllowance',
  DecreaseAllowanceAndCall = 'DecreaseAllowanceAndCall',
  DecreaseSupply = 'DecreaseSupply',
  DefaultOperators = 'DefaultOperators',
  Deployed = 'Deployed',
  GetApproved = 'GetApproved',
  GetCurrentFor = 'GetCurrentFor',
  Granularity = 'Granularity',
  HasHash = 'HasHash',
  HolderAt = 'HolderAt',
  HolderCount = 'HolderCount',
  IncreaseAllowance = 'IncreaseAllowance',
  IncreaseAllowanceAndCall = 'IncreaseAllowanceAndCall',
  IncreaseSupply = 'IncreaseSupply',
  IsApprovedForAll = 'IsApprovedForAll',
  IsHolder = 'IsHolder',
  IsPauser = 'IsPauser',
  IsOperatorFor = 'IsOperatorFor',
  IsOwner = 'IsOwner',
  IsSuperseded = 'IsSuperseded',
  IsVerified = 'IsVerified',
  Mint = 'Mint',
  MintWithTokenURI = 'MintWithTokenURI',
  Name = 'Name',
  NewWallet = 'NewWallet',
  OnErc721Received = 'OnErc721Received',
  OnErc1155BatchReceived = 'OnErc1155BatchReceived',
  OnErc1155Received = 'OnErc1155Received',
  OperatorBurn = 'OperatorBurn',
  OperatorSend = 'OperatorSend',
  Owner = 'Owner',
  OwnerOf = 'OwnerOf',
  Rate = 'Rate',
  RemoveVerified = 'RemoveVerified',
  RenounceOwnership = 'RenounceOwnership',
  RenouncePauser = 'RenouncePauser',
  RevokeOperator = 'RevokeOperator',
  SafeBatchTransferFrom = 'SafeBatchTransferFrom',
  SafeTransferFrom = 'SafeTransferFrom',
  SafeTransferFromData = 'SafeTransferFromData',
  SafeTransferFromValueData = 'SafeTransferFromValueData',
  Send = 'Send',
  SetApprovalForAll = 'SetApprovalForAll',
  SetRate = 'SetRate',
  SupportsInterface = 'SupportsInterface',
  Symbol = 'Symbol',
  Target = 'Target',
  TokensReceived = 'TokensReceived',
  TokensToSend = 'TokensToSend',
  TokenByIndex = 'TokenByIndex',
  TokenFallback = 'TokenFallback',
  TokenOfOwnerByIndex = 'TokenOfOwnerByIndex',
  TokenUri = 'TokenUri',
  TotalSupply = 'TotalSupply',
  Transfer = 'Transfer',
  TransferData = 'TransferData',
  TransferDataFallback = 'TransferDataFallback',
  TransferAndCall = 'TransferAndCall',
  TransferFrom = 'TransferFrom',
  TransferFromAndCall = 'TransferFromAndCall',
  TransferOwnership = 'TransferOwnership',
  UpdateVerified = 'UpdateVerified',
  URI = 'URI',
  Pause = 'Pause',
  Pin = 'Pin',
  Paused = 'Paused',
  Unpause = 'Unpause',
  Upgrade = 'Upgrade',
  Resume = 'Resume',
  Wallet = 'Wallet',
  Withdraw = 'Withdraw',
}

export enum ErcName {
  AllFunctions = 'AllFunctions',
  Isy20 = 'Isy20',
  Isy20Burnable = 'Isy20Burnable',
  Isy20Capped = 'Isy20Capped',
  Isy20Detailed = 'Isy20Detailed',
  Isy20Mintable = 'Isy20Mintable',
  Isy20Pausable = 'Isy20Pausable',
  Isy165 = 'Isy165',
  Isy721 = 'Isy721',
  Isy721Burnable = 'Isy721Burnable',
  Isy721Receiver = 'Isy721Receiver',
  Isy721Metadata = 'Isy721Metadata',
  Isy721Enumerable = 'Isy721Enumerable',
  Isy721Pausable = 'Isy721Pausable',
  Isy721Mintable = 'Isy721Mintable',
  Isy721MetadataMintable = 'Isy721MetadataMintable',
  Isy721Full = 'Isy721Full',
  Isy820 = 'Isy820',
  Isy1155 = 'Isy1155',
  Isy1155Receiver = 'Isy1155Receiver',
  Isy1155Metadata = 'Isy1155Metadata',
  Isy223 = 'Isy223',
  Isy223Receiver = 'Isy223Receiver',
  Isy621 = 'Isy621',
  Isy777 = 'Isy777',
  Isy777Receiver = 'Isy777Receiver',
  Isy777Sender = 'Isy777Sender',
  Isy827 = 'Isy827',
  Isy884 = 'Isy884',
  Upgradeable = 'Upgradeable',
  Ownable = 'Ownable',
  PauserRole = 'PauserRole',
  IsyFS = 'IsyFS',
}